+++ __tests__/AudioPlayer.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { fireEvent, render, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import '../pages/setupTest';
+import '../utils/setupTest';
 import AudioPlayer from '@/pages/AudioPlayer';
-import setUpMockHTMLMediaElement from '@/pages/setupTest';
+import setUpMockHTMLMediaElement from '@/utils/setupTest';
 
 const audio = new Audio(
   'https://p.scdn.co/mp3-preview/ec256975df2ce04185ba00f5f70a125cbcb4ae5e?cid=2847cfc683244615b79a93a6e24f375c'
